==> R/03_Swiss-SEP3.R <==
# starting from two csv files generated by
# analyses/SEP2_supplement.do 
# use R to generate Rds & shp files

library(readr)
library(dplyr)
library(sf)

ssep3_user <- read_csv("FINAL/CSV/ssep3_user_geo.csv", 
                       col_types = cols(gisid = col_integer(), 
                                        geox = col_integer(), geoy = col_integer(), 
                                        ssep2_t = col_integer(), 
                                        ssep2_q = col_integer(), 
                                        ssep2_d = col_integer(),
                                        ssep3_t = col_integer(), 
                                        ssep3_q = col_integer(), 
                                        ssep3_d = col_integer()
                                        )
                       ) 

# View(ssep3_user)

write_rds(ssep3_user, "FINAL/RDS/ssep3_user.Rds") 

# geo dataset 
ssep3_user_geo <- ssep3_user %>% 
  st_as_sf(coords = c("geox", "geoy"), 
           crs = 2056,
           remove = FALSE)

# plot(st_geometry(ssep3_user_geo))
# View(st_drop_geometry(ssep3_user_geo))

write_rds(ssep3_user_geo, "FINAL/RDS/ssep3_user_geo.Rds") 

# st_write(ssep3_user_geo, "FINAL/GEO/ssep3_user_geo.shp", delete_dsn = TRUE)
st_write(ssep3_user_geo, "FINAL/GEO/ssep3_user_geo.gpkg", delete_dsn = TRUE)
