==> R/04_sep-diff.Rmd <==
---
title: "Swiss-SEP 2.0"
description: "Differences 1.0 vs 2.0"
date: "`r Sys.Date()`"
author:
  - name: Radoslaw Panczak 
    url: https://github.com/RPanczak
    affiliation: ISPM
    affiliation_url: https://www.ispm.unibe.ch/
    orcid_id: 0000-0001-5141-683X
# bibliography: biblio.bib
output:
  distill::distill_article:
    highlight: pygments
    toc: true
    toc_depth: 1
    number_sections: true
editor_options: 
  chunk_output_type: console
---

<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, 
                      cache = FALSE, 
                      # fig.width=8, fig.height=6, 
                      # out.width="800px", out.height="600px",
                      dpi=300)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
options(scipen=999)
set.seed(12345)
library(pacman) 
p_load(readr, dplyr, ggplot2, scales, sf, tmap, tmaptools)
tmap_mode("view") # makes map interactive
```

<!-- ------------------------------------------------------------ --> 

# Data

## Input 

```{r load, include=FALSE}
# important to work with planar coords!

sep1 <- read_rds("data/Swiss-SEP1/ssep_user_geo.Rds") %>% 
  select(v0_buildid, ssep, ssep_t, ssep_q, ssep_d) %>% 
  rename(ssep1 = ssep,
         ssep1_t = ssep_t, 
         ssep1_q = ssep_q, 
         ssep1_d = ssep_d) %>% 
  mutate(ssep1_d = factor(ssep1_d,
                          levels = 1:10,
                          labels = c("1st - lowest", 
                                     "2", "3", "4", 
                                     "5th decile", 
                                     "6", "7", "8", "9", 
                                     "10th - highest")))

sep2 <- read_rds("data/Swiss-SEP2/ssep2_user_geo.Rds") %>% 
  select(gisid, ssep2, ssep2_t, ssep2_q, ssep2_d) %>% 
  mutate(ssep2_d = factor(ssep2_d,
                          levels = 1:10,
                          labels = c("1st - lowest", 
                                     "2", "3", "4", 
                                     "5th decile", 
                                     "6", "7", "8", "9", 
                                     "10th - highest")))
```

Using `r number(nrow(sep1), big.mark = ",")` n'hoods from version 1.0 and `r number(nrow(sep2), big.mark = ",")` from version 2.0.

```{r}
sep2 %>% 
  slice(1:5) %>% 
  knitr::kable(digits = c(0, 1, 0),
               caption = 'An example of Swiss-SEP 2.0 data.')
```

## Spatial join & distance

Calculations done with `sf` package funcion `st_join` (with `st_nearest_feature` option) and `st_distance` (of the nearest feature)

```{r eval=FALSE, include=FALSE}
# results with coordinates of 'sep2' features
# takes some time to compute distance

sep2_sep1_join <- st_join(sep2, sep1,
                          join = st_nearest_feature)

# adding distance
# solution from @Spacedman
# https://gis.stackexchange.com/questions/349955/getting-a-new-column-with-distance-to-the-nearest-feature-in-r
nearest <- st_nearest_feature(sep2, sep1)
distance <- st_distance(sep2, sep1[nearest, ], by_element = TRUE)
sep2_sep1_join$distance <- distance
rm(nearest, distance)

write_rds(sep2_sep1_join, "data/Swiss-SEP2/sep2_sep1_join.Rds")

sep2_sep1_join %>% 
  st_drop_geometry() %>% 
  haven::write_dta("data/Swiss-SEP2/sep2_sep1_join.dta")
```

<!-- ------------------------------------------------------------ --> 

# Distance discrepancies

Top 20 n'hood centres of 2.0 index (in red) that are furthest from 1.0 n'hood (in black). Using straight line distance.

```{r, layout="l-page", fig.width=6, fig.height=4}
sep2_sep1_join <- read_rds("data/Swiss-SEP2/sep2_sep1_join.Rds")

top20_dist <- sep2_sep1_join %>% 
  arrange(-distance) %>% 
  slice(1:20) %>% 
  mutate(rank = row_number(),
         distance = as.numeric(distance))

top20_dist_neigh <- sep1 %>% 
  inner_join(st_drop_geometry(select(top20_dist, v0_buildid)))

tm_shape(top20_dist) +
  tm_dots(col = "red") +
  tm_shape(top20_dist_neigh) +
  tm_dots()
```

<!-- ------------------------------------------------------------ --> 

# Exact values discrepancies

```{r}
p_load(blandr, BlandAltmanLeh)

bas_data <- sep2_sep1_join %>% 
  st_drop_geometry() %>% 
  # sample_frac(0.01) %>% 
  mutate(ssep1 = as.integer(ssep1),
         ssep2 = as.integer(ssep2))

blandr.output.text (bas_data$ssep1, bas_data$ssep2, sig.level=0.95 )
p_unload(blandr)
```

```{r echo=FALSE}
bas_list <- bland.altman.stats(bas_data$ssep1, bas_data$ssep2)
bas <- tibble(means = bas_list$means, diffs = bas_list$diffs)

# bas_list$mean.diffs 
# bas_list$critical.diff 

ggplot(bas, aes(x = means, y = diffs)) + 
  geom_count(alpha = 0.05) + 
  # geom_point() + 
  geom_hline(yintercept = bas_list$lines[-2], 
             linetype = 2, 
             size = 1,
             colour = "slateblue1") + 
  geom_hline(yintercept = bas_list$lines[2], 
             linetype = 2, 
             size = 1, 
             colour = "slateblue4") + 
  xlab("Mean of measurements") + 
  ylab("Difference") + 
  theme_classic()

p_unload(BlandAltmanLeh)
```

```{r eval=FALSE, include=FALSE}
# Diff histogram:
ggplot(bas, aes(x = diffs)) + 
  geom_histogram(binwidth = 1)
```

Mean diff and CI lines:

```{r}
bas_list$lines
```

Proportion within 95%:

```{r}
prop.table(table(bas_list$diffs >= bas_list$lines[1] & bas_list$diffs <= bas_list$lines[3]))
```
  
<!-- ------------------------------------------------------------ --> 

# Decile discrepancies

```{r}
start <- nrow(st_drop_geometry(sep2_sep1_join))

sep2_sep1_join %>% 
  st_drop_geometry() %>% 
  filter(ssep2_d == ssep1_d) %>% 
  nrow() -> agreement

sep2_sep1_join %>% 
  st_drop_geometry() %>% 
  mutate(diff = as.numeric(ssep2_d) - as.numeric(ssep1_d)) %>% 
  filter(diff >= -1 & diff <= 1) %>% 
  nrow() -> partial

sep2_sep1_join %>% 
  st_drop_geometry() %>% 
  mutate(diff = as.numeric(ssep2_d) - as.numeric(ssep1_d)) %>% 
  filter(diff == -9 | diff == 9) %>% 
  nrow() -> disagreement
```


Out of `r number(start, big.mark = ",")` n'hoods of SSEP 2.0 `r number(agreement, big.mark = ",")` or `r percent(agreement/start)` has been assigned to the same decile as **the closest** SSEP 1.0. 

If we relax this assumption and alow variation +/- 1 decile then `r percent(partial/start)` of n'hoods are matching.

There are only `r number(disagreement, big.mark = ",")` (`r percent(disagreement/start, accuracy = 0.001)`) n'hoods that are on the opposite scale of the index. 

## Full numbers

Counts

```{r}
sep_dif_xtab <- table(sep2_sep1_join$ssep1_d, sep2_sep1_join$ssep2_d) # first var in rows
# sep_dif_xtab

sep_dif_xtab %>% # 
  knitr::kable(caption = 'SSEP 1.0 (rows) vs. 2.0 (columns).')
```

Share (cell %)

```{r}
prop.table(sep_dif_xtab) %>% # cell percentages
  knitr::kable(digits = 3,
               caption = '% SSEP 1.0 (rows) vs. 2.0 (columns).')
```

```{r eval=FALSE, include=FALSE}
mosaicplot(sep_dif_xtab, shade=TRUE, legend=TRUE)

spineplot(sep2_sep1_join$ssep1_d, sep2_sep1_join$ssep2_d)

library(vcd)
agreementplot(sep_dif_xtab)
```

Visually 

```{r, layout="l-page", fig.width=6, fig.height=6}
sep_dif_freq <- as.data.frame(table(sep2_sep1_join$ssep1_d, sep2_sep1_join$ssep2_d)) %>% 
  rename(SSEP1 = Var1, SSEP2 = Var2)

sep_dif_freq %>% 
  ggplot(aes(SSEP1, Freq)) +
  geom_col() +
  facet_grid(vars(SSEP2))
```

Alternative view

```{r, layout="l-body-outset"}
# sep_dif_freq %>% 
#   ggplot(aes(SSEP1, SSEP2)) +
#   geom_tile(aes(fill = Freq / nrow(sep2_sep1_join)), colour = "black") +
#   scale_fill_distiller(palette = "RdYlBu", direction = 1)
# 
# sep_dif_freq %>% 
#   filter(SSEP1 != SSEP2) %>% 
#   ggplot(aes(SSEP1, SSEP2)) +
#   geom_tile(aes(fill = Freq / nrow(sep2_sep1_join)), colour = "black") +
#   scale_fill_distiller(palette = "RdYlBu", direction = 1)

sep_dif_freq %>% 
  ggplot(aes(SSEP1, SSEP2)) +
  geom_point(aes(size = Freq, color = Freq / nrow(sep2_sep1_join)), shape = 15) +
  scale_size_continuous(range = c(2, 10)) + 
  scale_color_distiller(palette = "RdYlBu", direction = 1) +
  theme_classic()
```

Excluding diagonal

```{r, layout="l-body-outset"}
sep_dif_freq %>% 
  filter(SSEP1 != SSEP2) %>% 
  ggplot(aes(SSEP1, SSEP2)) +
  geom_point(aes(size = Freq, color = Freq / nrow(sep2_sep1_join)), shape = 15) +
  scale_size_continuous(range = c(2, 10)) + 
  scale_color_distiller(palette = "RdYlBu", direction = 1) +
  theme_classic()
```

<!-- ------------------------------------------------------------ --> 

# Spatial discrepancies

## SSEP 2.0 higher

```{r, layout="l-page"}
diff_plus <- sep2_sep1_join %>% 
  mutate(diff = as.numeric(ssep2_d) - as.numeric(ssep1_d)) %>% 
  filter(diff == 9) %>% 
  select(-diff)

diff_plus_neigh <- sep1 %>% 
  inner_join(st_drop_geometry(select(diff_plus, v0_buildid)))

tm_shape(diff_plus) +
  tm_dots(col = "red") +
  tm_shape(diff_plus_neigh) +
  tm_dots()

# library(mapview)
# library(leaflet)
# library(leafpop)
# 
# mapview(list(diff_plus, diff_plus_neigh),
#         col.regions = list("red", "black"),
#         popup = list(popupTable(diff_plus,
#                            zcol = "ssep2"), 
#                      (popupTable(diff_plus_neigh,
#                            zcol = "ssep1")
#                      )))
```

Example of [redeveloped area](https://goo.gl/maps/VLQShvxNNjmCpPam9) that migh experienced a bump.

## SSEP 2.0 lower

```{r, layout="l-page"}
diff_minus <- sep2_sep1_join %>% 
  mutate(diff = as.numeric(ssep2_d) - as.numeric(ssep1_d)) %>% 
  filter(diff == -9) %>% 
  select(-diff)

diff_minus_neigh <- sep1 %>% 
  inner_join(st_drop_geometry(select(diff_minus, v0_buildid)))

tm_shape(diff_minus) +
  tm_dots(col = "red") +
  tm_shape(diff_minus_neigh) +
  tm_dots()

# mapview(list(diff_minus, diff_minus_neigh),
#         col.regions = list("red", "black"),
#         popup = list(popupTable(diff_minus,
#                            zcol = "ssep2"), 
#                      (popupTable(diff_minus_neigh,
#                            zcol = "ssep1")
#                      )))
```
