==> R/22_grid.Rmd <==
---
title: "Swiss-SEP project"
description: "Aggregating point data to grid"
date: "`r Sys.Date()`"
author:
  - name: Radoslaw Panczak 
    url: https://github.com/RPanczak
    affiliation: ISPM
    affiliation_url: https://www.ispm.unibe.ch/
    orcid_id: 0000-0001-5141-683X
output:
  distill::distill_article:
    highlight: pygments
    toc: true
    toc_depth: 1
    number_sections: true
editor_options: 
  chunk_output_type: console
---

<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, 
                      # fig.width=8, fig.height=6, 
                      # out.width="800px", out.height="600px",
                      dpi=300)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
options(scipen=999)
set.seed(12345)
library(pacman) 
p_load(tidyverse,
       sf, tmap,
       ggpubr)

# tmap_mode("plot")
import::from("sjmisc", "frq")
```

# Ancillary data

```{r}
canton <- st_read("../ISPM_geo/data-raw/BfS/ag-b-00.03-875-gg20/ggg_2020-LV95/shp/g1k20.shp") %>% 
  st_transform(crs = 2056) %>% 
  dplyr::select(KTNR, KTNAME)

lakes <- st_read("../ISPM_geo/data-raw/BfS/2019_THK_PRO/PRO/00_TOPO/K4_seenyyyymmdd/k4seenyyyymmdd_ch2007Poly.shp") %>% 
  st_transform(crs = 2056)

# rivers <- st_read("../ISPM_geo/data-raw/BfS/2019_THK_PRO/PRO/00_TOPO/K4_flusyyyymmdd/k4flusyyyymmdd_ch2007.shp") %>% 
#   st_transform(crs = 2056)

p_load(raster)
gc()
# rgdal::GDALinfo("../ISPM_geo/data-raw/swisstopo/SMR1000_2019/SMR1000_EE/SMR1000_GRELI.tif")
relief <- raster("../ISPM_geo/data-raw/swisstopo/SMR1000_2019/SMR1000_EE/SMR1000_GRELI.tif") %>% 
  mask(canton) %>% 
  as("SpatialPixelsDataFrame") %>%
  as.data.frame() %>%
  rename(value = SMR1000_GRELI)
p_unload(raster)
gc()

city_lines <- st_read("data/cities.shp") %>%
  st_transform(crs = 2056)

city_labels <- city_lines %>% 
  st_cast("POINT") %>% 
  group_by(name) %>%
  slice(seq(2, n(), by = 2)) %>% 
  ungroup()

city_points <- city_lines %>% 
  st_cast("POINT") %>% 
  group_by(name) %>%
  slice(seq(1, n(), by = 2)) %>% 
  ungroup()
```

# SEP

```{r}
sep3 <- read_rds("FINAL/RDS/ssep3_user_geo.Rds") %>% 
  select(gisid, 
         ssep2, ssep2_d,
         ssep3, ssep3_d)
```

# STATPOP offset

Used to define `offset` for `st_make_grid`

```{r eval=FALSE}
offset_bfs <- read_delim("data-raw/BfS/ag-b-00.03-vz2020statpop/STATPOP2020.zip", 
                         delim = ";", escape_double = FALSE, trim_ws = TRUE)[] %>% 
  mutate_all(as.integer) %>%  
  summarise(min_x = min(X_KOORD),
            min_y = min(Y_KOORD)) %>% 
  st_as_sf(coords = c("min_x", "min_y"), 
           crs = 21781,
           remove = FALSE)

write_rds(offset_bfs, "data/grid/offset_bfs.Rds")
offset_bfs <- read_rds("data/grid/offset_bfs.Rds")
```

# Generate grid

With lower left corner defined using offset from above  
Also, adding ID for each cell (based on row number).   

## 100m grid  

```{r eval=FALSE}
gc()

country_hex_100 <- canton %>% 
  st_make_grid(cellsize = 100, 
               offset = c(offset_bfs$min_x, offset_bfs$min_y),
               square = FALSE) %>% 
  st_sf() %>%
  st_cast("POLYGON") %>% 
  st_filter(canton, join = st_covers) %>%
  mutate(ID1 = row_number()) %>% 
  relocate(ID1)

write_rds(country_hex_100, "data/grid/country_hex_100.Rds")

rm(offset_bfs)
```

### Summarize per cell

```{r eval=FALSE}
country_hex_100 <- read_rds("data/grid/country_hex_100.Rds")
gc()

sep3_grid_hex_100 <- 
  st_join(sep3, country_hex_100, join = st_intersects) %>% 
  st_drop_geometry() %>% 
  group_by(ID1) %>% 
  summarise(n = n(),
            median_d = median(ssep3_d), 
            median = median(ssep3)) %>% 
  ungroup() %>% 
  mutate(median_d = factor(round(median_d),
                           levels = 1:10,
                           labels = c("1st - lowest",
                                      "2", "3", "4",
                                      "5th decile",
                                      "6", "7", "8", "9",
                                      "10th - highest")))

# frq(sep3_grid$median_d)

sep3_grid_hex_100_sf <- country_hex_100 %>% 
  inner_join(sep3_grid_hex_100)

write_rds(sep3_grid_hex_100_sf, "data/grid/sep3_grid_hex_100_sf.Rds")

# plot(sep3_grid_sf[, 3])

sep3_grid_hex_100_clip_sf <- sep3_grid_hex_100_sf %>% 
  st_intersection(st_union(canton)) %>% 
  st_collection_extract("POLYGON") %>% 
  st_difference(st_union(lakes)) %>% 
  st_collection_extract("POLYGON")

write_rds(sep3_grid_hex_100_clip_sf, "data/grid/sep3_grid_hex_100_clip_sf.Rds")

rm(country_hex_100, sep3_grid_hex_100); gc()
```

```{r}
sep3_grid_hex_100_sf <- read_rds("data/grid/sep3_grid_hex_100_sf.Rds")
# sep3_grid_hex_100_clip_sf <- read_rds("data/grid/sep3_grid_hex_100_clip_sf.Rds")
```

#### Plot

```{r}
ggplot() + 
  geom_sf(
    data = sep3_grid_hex_100_sf,
    # data = sep3_grid_hex_100_clip_sf,
    aes(fill = median_d),
    alpha = 0.66, color = NA) +
  geom_sf(data = lakes, color = NA,  fill = rgb(156, 213, 248, max = 255)) + 
  geom_sf(data = canton, fill = NA, color = gray(.5), size = 0.5) + 
  scale_fill_brewer(palette = "RdYlGn") +
  theme_void() + 
  theme(legend.position = c(.05, .95), 
        legend.justification = c("left", "top"),
        legend.direction = "vertical",
        plot.margin = unit(c(-5, 0, -5, -5), "mm")) +
  guides(fill = guide_legend(title = "Swiss-SEP 3 index",
                             override.aes = list(alpha = 1)))

ggsave("carto/07_sep3-grid_hex_100.png", 
       width = 297, height = 210, units = "mm", dpi = 300,
       bg = "white")
```

## 500m grid  

```{r eval=FALSE}
gc()

country_hex_500 <- canton %>% 
  st_make_grid(cellsize = 500, 
               offset = c(offset_bfs$min_x, offset_bfs$min_y),
               square = FALSE) %>% 
  st_sf() %>%
  st_cast("POLYGON") %>% 
  st_filter(canton, join = st_covers) %>%
  mutate(ID1 = row_number()) %>% 
  relocate(ID1)

write_rds(country_hex_500, "data/grid/country_hex_500.Rds")

rm(offset_bfs)
```

### Summarize per cell

```{r eval=FALSE}
country_hex_500 <- read_rds("data/grid/country_hex_500.Rds")
gc()

sep3_grid_hex_500 <- 
  st_join(sep3, country_hex_500, join = st_intersects) %>% 
  st_drop_geometry() %>% 
  group_by(ID1) %>% 
  summarise(n = n(),
            median2_d = median(ssep2_d), 
            median2 = median(ssep2),
            median3_d = median(ssep3_d), 
            median3 = median(ssep3),
            median_diff = round(median3_d) - round(median2_d)
  ) %>% 
  ungroup() %>% 
  mutate(median2_d = factor(round(median2_d),
                            levels = 1:10,
                            labels = c("1st - lowest",
                                       "2", "3", "4",
                                       "5th decile",
                                       "6", "7", "8", "9",
                                       "10th - highest")),
         median3_d = factor(round(median3_d),
                            levels = 1:10,
                            labels = c("1st - lowest",
                                       "2", "3", "4",
                                       "5th decile",
                                       "6", "7", "8", "9",
                                       "10th - highest")))

# frq(sep3_grid_hex_500$median_d)
# frq(sep3_grid_hex_500$median_diff)

sep3_grid_hex_500_sf <- country_hex_500 %>% 
  inner_join(sep3_grid_hex_500)

write_rds(sep3_grid_hex_500_sf, "data/grid/sep3_grid_hex_500_sf.Rds")

# plot(sep3_grid_sf[, 3])

sep3_grid_hex_500_clip_sf <- sep3_grid_hex_500_sf %>% 
  st_intersection(st_union(canton)) %>% 
  st_collection_extract("POLYGON") %>% 
  st_difference(st_union(lakes)) %>% 
  st_collection_extract("POLYGON")

write_rds(sep3_grid_hex_500_clip_sf, "data/grid/sep3_grid_hex_500_clip_sf.Rds")

rm(country_hex_500, sep3_grid_hex_500); gc()
```

```{r}
sep3_grid_hex_500_sf <- read_rds("data/grid/sep3_grid_hex_500_sf.Rds")
# sep3_grid_hex_500_clip_sf <- read_rds("data/grid/sep3_grid_hex_500_clip_sf.Rds")
```

### Plot deciles

```{r}
median2_d <- 
  ggplot() + 
  
  geom_raster(
    data = relief,
    inherit.aes = FALSE,
    aes(
      x = x,
      y = y,
      alpha = value
    )
  ) +
  scale_alpha(name = "",
              range = c(0.4, 0),
              guide = "none") + 
  
  geom_sf(
    data = sep3_grid_hex_500_sf,
    aes(fill = median2_d),
    alpha = 0.66, color = NA,
    show.legend = TRUE) +
  scale_fill_brewer(palette = "RdYlGn") +
  geom_sf(data = lakes, 
          color = NA, fill = rgb(156, 213, 248, max = 255),
          show.legend = FALSE) + 
  geom_sf(data = canton, 
          fill = NA, color = "black", size = 0.2,
          show.legend = FALSE) + 
  
  geom_sf(data = city_lines,
          show.legend = FALSE,
          inherit.aes = FALSE) +
  
  geom_sf(data = city_points,
          size = 1,
          show.legend = FALSE,
          inherit.aes = FALSE) +
  
  geom_sf_label(data = city_labels,
                label.size  = NA,
                aes(label = name),
                show.legend = FALSE,
                inherit.aes = FALSE) +
  
  theme_void() + 
  theme(legend.position = c(.03, 1), 
        legend.justification = c("left", "top"),
        legend.direction = "vertical",
        plot.margin = unit(c(-10, -10, -10, -5), "mm"),
        legend.title = element_blank(),
        plot.title = element_text(hjust = 0.5, vjust = -5)) +
  guides(fill = guide_legend(override.aes = list(alpha = 1))) + 
  ggtitle("Swiss-SEP 2 index")


median3_d <- 
  ggplot() + 
  
  geom_raster(
    data = relief,
    inherit.aes = FALSE,
    aes(
      x = x,
      y = y,
      alpha = value
    )
  ) +
  scale_alpha(name = "",
              range = c(0.4, 0),
              guide = "none") + 
  
  geom_sf(
    data = sep3_grid_hex_500_sf, 
    aes(fill = median3_d),
    alpha = 0.66, color = NA,
    show.legend = FALSE) +
  scale_fill_brewer(palette = "RdYlGn", guide = "none") +
  geom_sf(data = lakes, 
          color = NA, fill = rgb(156, 213, 248, max = 255),
          aes(shape = factor(1, labels = "Lakes")),
          show.legend = TRUE) + 
  geom_sf(data = canton, 
          aes(color = factor(1, labels = "Canton borders")),
          fill = NA, size = 0.2,
          show.legend = TRUE) +  
  scale_color_manual(values = "black") +
  
  geom_sf(data = city_lines,
          show.legend = FALSE,
          inherit.aes = FALSE) +
  
  geom_sf(data = city_points,
          size = 1,
          show.legend = FALSE,
          inherit.aes = FALSE) +
  
  geom_sf_label(data = city_labels,
                label.size  = NA,
                aes(label = name),
                show.legend = FALSE,
                inherit.aes = FALSE) +
  
  theme_void() + 
  theme(legend.position = c(.03, 1), 
        legend.justification = c("left", "top"),
        legend.direction = "vertical",
        plot.margin = unit(c(-10, -10, -10, -5), "mm"),
        legend.title = element_blank(),
        plot.title = element_text(hjust = 0.5, vjust = -5)) +
  guides(shape = guide_legend(override.aes = list(size = 1,
                                                  color = NA)),
         color = guide_legend(override.aes = list(fill = NA, 
                                                  size = 1,
                                                  color = "black"))) + 
  ggtitle("Swiss-SEP 3 index")

ggarrange(
  median2_d, median3_d, 
  ncol = 1
)

ggsave("analyses/Figure_1.png", 
       width = 210, height = 297, units = "mm", dpi = 300,
       bg = "white")
```

### Plot diff

```{r}
# we need more colours for differences on deciles
p_load(RColorBrewer)

frq(sep3_grid_hex_500_sf$median_diff)
length(unique(sep3_grid_hex_500_sf$median_diff))

# stretch the pallete from 11 to 19
brewer <- brewer.pal(11, "RdYlGn")
cols <- c(colorRampPalette(c(brewer[1], brewer[6]))(10), 
          colorRampPalette(c(brewer[6], brewer[11]))(10)[-1])
length(unique(cols))

barplot(1:19, col = cols)

# masking middle with light grey? get rids of yellow middle band
cols[9]  <- "#E5E5E5"
cols[10] <- "#E5E5E5"
cols[11] <- "#E5E5E5"

barplot(1:19, col = cols)
```

```{r}
ggplot() + 
  
  # geom_raster(
  #   data = relief,
  #   inherit.aes = FALSE,
  #   aes(
  #     x = x,
  #     y = y,
  #     alpha = value
  #   )
  # ) +
  # scale_alpha(name = "",
  #             range = c(0.4, 0),
  #             guide = "none") + 
  
  geom_sf(
    data = sep3_grid_hex_500_sf,
    aes(fill = factor(median_diff)),
    alpha = 0.66, color = NA,
    show.legend = TRUE) +
  scale_fill_manual(values = cols) +
  geom_sf(data = lakes, 
          color = NA, fill = rgb(156, 213, 248, max = 255),
          show.legend = FALSE) + 
  geom_sf(data = canton, 
          fill = NA, color = "black", size = 0.2,
          show.legend = FALSE) + 
  
  geom_sf(data = city_lines,
          show.legend = FALSE,
          inherit.aes = FALSE) +
  
  geom_sf(data = city_points,
          size = 1,
          show.legend = FALSE,
          inherit.aes = FALSE) +
  
  geom_sf_label(data = city_labels,
                label.size  = NA,
                aes(label = name),
                show.legend = FALSE,
                inherit.aes = FALSE) +
  
  theme_void() + 
  theme(legend.position = c(.03, 1), 
        legend.justification = c("left", "top"),
        legend.direction = "vertical",
        plot.margin = unit(c(-10, -10, -10, -5), "mm"),
        legend.title = element_blank(),
        plot.title = element_text(hjust = 0.5, vjust = -5)) +
  guides(fill = guide_legend(override.aes = list(alpha = 1))) + 
  ggtitle("Swiss-SEP 3 - 2 difference")

ggsave("carto/08_sep-diff-grid_hex_500.png", 
       height = 210, width = 297, units = "mm", dpi = 300,
       bg = "white")
```


# Making a raster

```{r}
rm(sep3); gc()
```
